+++ fastly/ngwaf/v1/workspaces/alerts/doc.go
@@ -1,0 +1,3 @@
+// Package alerts contains API operations to create, delete
+// get,  list, rotate and update Fastly Next-Gen WAF workspace alerts.
+package alerts