+++ fastly/ngwaf/v1/workspaces/virtualpatches/api_update.go
@@ -11,6 +11,10 @@ import (
 // UpdateInput specifies the information needed for the Update function
 // to perform the operation.
 type UpdateInput struct {
+	// Action is the action to take when signal for virtual patch is detected.
+	Action *string `json:"action"`
+	// Context, if supplied, will be used as the Request's context.
+	Context *context.Context `json:"-"`
 	// Enabled is the toggle status indicator of the VirtualPatch.
 	Enabled *bool `json:"enabled"`
 	// Mode is action to take when a signal for virtual patch is